hypergraph -> simplicial
