+++ configs/transforms/liftings/graph2simplicial/line_lifting.yaml
@@ -1,3 +1,3 @@
 transform_type: "lifting"
-transform_name: "HypergraphLineLifting"
+transform_name: "SimplicialLineLifting"
 feature_lifting: ProjectionSum
