+++ modules/transforms/data_transform.py
@@ -12,19 +12,19 @@ from modules.transforms.liftings.graph2cell.cycle_lifting import CellCycleLiftin
 from modules.transforms.liftings.graph2hypergraph.knn_lifting import (
     HypergraphKNNLifting,
 )
-from modules.transforms.liftings.graph2hypergraph.line_lifting import (
-    HypergraphLineLifting,
-)
 from modules.transforms.liftings.graph2simplicial.clique_lifting import (
     SimplicialCliqueLifting,
+)
+from modules.transforms.liftings.graph2simplicial.line_lifting import (
+    SimplicialLineLifting,
 )
 
 TRANSFORMS = {
     # Graph -> Hypergraph
     "HypergraphKNNLifting": HypergraphKNNLifting,
-    "HypergraphLineLifting": HypergraphLineLifting,
     # Graph -> Simplicial Complex
     "SimplicialCliqueLifting": SimplicialCliqueLifting,
+    "SimplicialLineLifting": SimplicialLineLifting,
     # Graph -> Cell Complex
     "CellCycleLifting": CellCycleLifting,
     # Feature Liftings
